system/readline: Make it work without CONFIG_FILE_STREAM
